pyproject.toml
[tool.poetry]
name = "nl2bench"
version = "0.4.0"
description = "Converts from combinational netlists to the BENCH format for DFT"
authors = ["Mohamed Gaber <donn@efabless.com>"]
license = "Apache-2.0"
readme = "Readme.md"
include = ["_nl2bench_antlr4_liblogic/*.py", "_nl2bench_antlr4_verilog/*.py"]

[tool.poetry.dependencies]
python = ">=3.8,<4"
click = ">=8,<9"
antlr4-python3-runtime = "==4.9.*"
libparse = "^0.3.1"
frozendict = "^2.4.4"


[tool.poetry.dev-dependencies]
wheel = "*"
black = ">=24.4.0,<25"

flake8 = ">=4"
flake8-pytest-style = "*"

pytest = "*"
pytest-xdist = "*"
pytest-cov = "*"

[tool.poetry.scripts]
nl2bench = "nl2bench.__main__:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
